# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Rumah Sakit",
        "description": "API ini digunakan untuk mengelola data pasien, dokter, konsultasi, dan lainnya.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server API Lokal"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/dokter": {
            "get": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Menampilkan semua dokter",
                "operationId": "56513231100bce5b3fbc8d8665026b73",
                "responses": {
                    "200": {
                        "description": "Daftar dokter berhasil diambil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Menambahkan dokter baru",
                "operationId": "0ed4d306239fbc82987eb608fd5b7c32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "spesialisasi",
                                    "no_telepon",
                                    "email"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "spesialisasi": {
                                        "type": "string"
                                    },
                                    "no_telepon": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dokter berhasil ditambahkan"
                    }
                }
            }
        },
        "/api/dokter/{id}": {
            "get": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Menampilkan detail dokter",
                "operationId": "8a5c11bbb78172077f0c33f66c455a8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dokter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data dokter berhasil diambil"
                    },
                    "404": {
                        "description": "Dokter tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Memperbarui data dokter",
                "operationId": "3c37ae6731c0581fcf8b24593d50b40c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dokter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "spesialisasi": {
                                        "type": "string"
                                    },
                                    "no_telepon": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data dokter berhasil diperbarui"
                    },
                    "404": {
                        "description": "Dokter tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dokter"
                ],
                "summary": "Menghapus dokter",
                "operationId": "a369d595c876867767ab1ffd27c14f53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dokter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dokter berhasil dihapus"
                    },
                    "404": {
                        "description": "Dokter tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}